+++ lib/ts/utils/sessionClaimValidatorStore.ts
@@ -1,4 +1,3 @@
-"use strict";
 /* Copyright (c) 2020, VRAI Labs and/or its affiliates. All rights reserved.
  *
  * This software is licensed under the Apache License, Version 2.0 (the
@@ -13,11 +12,17 @@
  * License for the specific language governing permissions and limitations
  * under the License.
  */
-Object.defineProperty(exports, "__esModule", { value: true });
-var SessionClaimValidator = /** @class */ (function() {
-    function SessionClaimValidator(id) {
-        this.id = id;
-    }
-    return SessionClaimValidator;
-})();
-exports.SessionClaimValidator = SessionClaimValidator;
+
+import { SessionClaimValidator } from "../types";
+
+export class SessionClaimValidatorStore {
+    private static claimValidatorsAddedByOtherRecipes: SessionClaimValidator[] = [];
+
+    static addClaimValidatorFromOtherRecipe = (builder: SessionClaimValidator) => {
+        SessionClaimValidatorStore.claimValidatorsAddedByOtherRecipes.push(builder);
+    };
+
+    static getClaimValidatorsAddedByOtherRecipes = (): SessionClaimValidator[] => {
+        return SessionClaimValidatorStore.claimValidatorsAddedByOtherRecipes;
+    };
+}