+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Added
+
+- Added the `API_INVALID_CLAIM` event and the related `invalidClaimStatusCode` config prop.
+- Adds base classes, types for session claims, and a `validateClaims` functions.
+
 ## [13.0.1] - 2022-06-27
 - Fixes node dependency for integration tests